usuarios modificado

+++ almacen/index.php
@@ -196,8 +196,12 @@ include '../app/controllers/almacen/listado_de_productos.php';
                                                     <center>
                                                         <div class="btn-group">
                                                             <a href="show.php?id=<?php echo $id_producto; ?>" type="button" class="btn btn-info btn-sm"><i class="fa fa-eye"></i>Ver</a>
-                                                            <a href="update.php?id=<?php echo $id_producto; ?>" type="button" class="btn btn-success btn-sm"><i class="fa fa-pencil-alt"></i>Editar</a>
-                                                            <a href="delete.php?id=<?php echo $id_producto; ?>" type="button" class="btn btn-danger btn-sm"><i class="fa fa-trash"></i>Borrar</a>
+                                                            <?php if ($rol_sesion == "Administrador" || $rol_sesion == "Encargado de compras") { ?>
+                                                                <a href="update.php?id=<?php echo $id_producto; ?>" type="button" class="btn btn-success btn-sm"><i class="fa fa-pencil-alt"></i>Editar</a>
+                                                            <?php } ?>
+                                                            <?php if ($rol_sesion == "Administrador") { ?>
+                                                                <a href="delete.php?id=<?php echo $id_producto; ?>" type="button" class="btn btn-danger btn-sm"><i class="fa fa-trash"></i>Borrar</a>
+                                                            <?php } ?>
                                                         </div>
                                                     </center>
                                                 </td>
